refactor(tests): tweak test_cursor to make it more similar to the async version
