+++ tests/test_cursor_async.py
@@ -1,17 +1,24 @@
-import pytest
+"""
+Tests common to psycopg.AsyncCursor and its subclasses.
+"""
+
 import weakref
 import datetime as dt
-from typing import List
+from typing import Any, List
+from contextlib import aclosing
+
+import pytest
 
 import psycopg
 from psycopg import sql, rows
 from psycopg.adapt import PyFormat
 from psycopg.types import TypeInfo
 
-from .utils import alist, gc_collect, raiseif
-from .test_cursor import my_row_factory, ph
-from .test_cursor import execmany, _execmany  # noqa: F401
+from .utils import alist, anext
+from .utils import gc_collect, raiseif
 from .fix_crdb import crdb_encoding
+from ._test_cursor import my_row_factory, ph
+from ._test_cursor import execmany, _execmany  # noqa: F401
 
 execmany = execmany  # avoid F811 underneath
 
@@ -19,7 +26,7 @@ execmany = execmany  # avoid F811 underneath
 @pytest.fixture(
     params=[psycopg.AsyncCursor, psycopg.AsyncClientCursor, psycopg.AsyncRawCursor]
 )
-def aconn(aconn, request, anyio_backend):
+async def aconn(aconn, request, anyio_backend):
     aconn.cursor_factory = request.param
     return aconn
 
@@ -144,6 +151,42 @@ async def test_execute_sql(aconn):
     cur = aconn.cursor()
     await cur.execute(sql.SQL("select {value}").format(value="hello"))
     assert (await cur.fetchone()) == ("hello",)
+
+
+async def test_query_parse_cache_size(aconn):
+    cur = aconn.cursor()
+    cls = type(cur)
+
+    # Warning: testing internal structures. Test might need refactoring with the code.
+    cache: Any
+    if cls is psycopg.AsyncCursor:
+        cache = psycopg._queries._query2pg
+    elif cls is psycopg.AsyncClientCursor:
+        cache = psycopg._queries._query2pg_client
+    elif cls is psycopg.AsyncRawCursor:
+        pytest.skip("RawCursor has no query parse cache")
+    else:
+        assert False, cls
+
+    cache.cache_clear()
+    ci = cache.cache_info()
+    h0, m0 = ci.hits, ci.misses
+    tests = [
+        (f"select 1 -- {'x' * 3500}", (), h0, m0 + 1),
+        (f"select 1 -- {'x' * 3500}", (), h0 + 1, m0 + 1),
+        (f"select 1 -- {'x' * 4500}", (), h0 + 1, m0 + 1),
+        (f"select 1 -- {'x' * 4500}", (), h0 + 1, m0 + 1),
+        (f"select 1 -- {'%s' * 40}", ("x",) * 40, h0 + 1, m0 + 2),
+        (f"select 1 -- {'%s' * 40}", ("x",) * 40, h0 + 2, m0 + 2),
+        (f"select 1 -- {'%s' * 60}", ("x",) * 60, h0 + 2, m0 + 2),
+        (f"select 1 -- {'%s' * 60}", ("x",) * 60, h0 + 2, m0 + 2),
+    ]
+    for i, (query, params, hits, misses) in enumerate(tests):
+        pq = cur._query_cls(psycopg.adapt.Transformer())
+        pq.convert(query, params)
+        ci = cache.cache_info()
+        assert ci.hits == hits, f"at {i}"
+        assert ci.misses == misses, f"at {i}"
 
 
 async def test_execute_many_results(aconn):
@@ -459,13 +502,12 @@ async def test_rownumber_none(aconn, query):
 
 async def test_rownumber_mixed(aconn):
     cur = aconn.cursor()
-    await cur.execute(
-        """
-select x from generate_series(1, 3) x;
-set timezone to utc;
-select x from generate_series(4, 6) x;
-"""
-    )
+    queries = [
+        "select x from generate_series(1, 3) x",
+        "set timezone to utc",
+        "select x from generate_series(4, 6) x",
+    ]
+    await cur.execute(";\n".join(queries))
     assert cur.rownumber == 0
     assert await cur.fetchone() == (1,)
     assert cur.rownumber == 1
@@ -497,8 +539,7 @@ async def test_iter_stop(aconn):
         break
 
     assert (await cur.fetchone()) == (3,)
-    async for rec in cur:
-        assert False
+    assert (await alist(cur)) == []
 
 
 async def test_row_factory(aconn):
@@ -635,22 +676,22 @@ async def test_stream_sql(aconn):
 
 async def test_stream_row_factory(aconn):
     cur = aconn.cursor(row_factory=rows.dict_row)
-    ait = cur.stream("select generate_series(1,2) as a")
-    assert (await ait.__anext__())["a"] == 1
+    it = cur.stream("select generate_series(1,2) as a")
+    assert (await anext(it))["a"] == 1
     cur.row_factory = rows.namedtuple_row
-    assert (await ait.__anext__()).a == 2
+    assert (await anext(it)).a == 2
 
 
 async def test_stream_no_row(aconn):
     cur = aconn.cursor()
-    recs = [rec async for rec in cur.stream("select generate_series(2,1) as a")]
+    recs = await alist(cur.stream("select generate_series(2,1) as a"))
     assert recs == []
 
 
 @pytest.mark.crdb_skip("no col query")
 async def test_stream_no_col(aconn):
     cur = aconn.cursor()
-    recs = [rec async for rec in cur.stream("select")]
+    recs = await alist(cur.stream("select"))
     assert recs == [()]
 
 
@@ -689,11 +730,9 @@ async def test_stream_error_notx(aconn):
 async def test_stream_error_python_to_consume(aconn):
     cur = aconn.cursor()
     with pytest.raises(ZeroDivisionError):
-        gen = cur.stream("select generate_series(1, 10000)")
-        async for rec in gen:
-            1 / 0
-
-    await gen.aclose()
+        async with aclosing(cur.stream("select generate_series(1, 10000)")) as gen:
+            async for rec in gen:
+                1 / 0
     assert aconn.info.transaction_status in (
         aconn.TransactionStatus.INTRANS,
         aconn.TransactionStatus.INERROR,
